fix(xray): Protector.protect fd как Long (gomobile Go int → Java long)

+++ app/src/main/java/pw/x4/ninety/vpn/XrayController.kt
@@ -29,7 +29,8 @@ object XrayController {
         stop()
         try {
             Xraybridge.registerDialerController(object : Protector {
-                override fun protect(fd: Int): Boolean = service.protect(fd)
+                // gomobile биндит Go `int` как Java `long` → fd:Long (VpnService.protect ждёт Int).
+                override fun protect(fd: Long): Boolean = service.protect(fd.toInt())
             })
         } catch (_: Throwable) {}
         Xraybridge.runXrayFromJSON(buildConfig(bridges)) // бросит при ошибке конфига
